divide note and note overlay into components
