+++ src/components/index.jsx
@@ -3,3 +3,9 @@ export {NavBar}
 
 import {SideNav} from "./SideNav/sideNav";
 export{SideNav}
+
+import {Note} from "./Note/note"
+export{Note}
+
+import {NoteOverlay} from "./noteOverlay/noteOverlay"
+export{NoteOverlay}
